requirements.txt
flask==2.0.1
flask-restful==0.3.10
mysql-connector-python==8.0.26
python-dotenv==1.0.1
confluent-kafka==1.8.2
numpy==1.21.2
scikit-learn==0.24.2
joblib==1.0.1
python-dateutil==2.8.2 